Move nowarn test to warns, update check
